requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Automat==20.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bitarray==1.2.2
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cymem==2.0.5
cytoolz==0.11.0
Deprecated==1.2.12
dnspython==2.1.0
email-validator==1.1.3
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0/en_core_web_lg-3.0.0-py3-none-any.whl
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gensim==4.0.1
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.1.0
gspread==3.7.0
h2==3.2.0
hexbytes==0.2.1
hpack==3.0.0
httplib2==0.19.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipfshttpclient==0.7.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
lizard==1.17.7
llvmlite==0.36.0
lru-dict==1.1.7
lxml==4.6.3
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.1.0
murmurhash==1.0.5
netaddr==0.8.0
nltk==3.6.2
numba==0.53.1
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
parsel==1.6.0
parsimonious==0.8.1
pathy==0.5.2
preshed==3.0.5
priority==1.3.0
Protego==0.1.16
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.4
PyDispatcher==2.0.5
pyenchant==3.2.0
PyGithub==1.55
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyodbc==4.0.31
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.18.0
pyspellchecker==0.6.2
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.6.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rlp==2.0.1
rsa==4.7.2
scipy==1.6.3
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
smart-open==3.0.0
smmap==4.0.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.4.18
srsly==2.4.1
thinc==8.0.3
tika==1.24
toolz==0.11.1
tqdm==4.61.0
Twisted==21.2.0
typer==0.3.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
varint==1.0.2
w3lib==1.22.0
wasabi==0.8.2
web3==5.22.0
websockets==9.1
Werkzeug==2.0.1
Whoosh==2.7.4
wrapt==1.12.1
WTForms==2.3.3
yarl==1.6.3
zope.interface==5.4.0

locust~=2.1.0